main.py
"""
personAI - Personal AI Assistant
Main application entry point
"""
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="personAI",
    description="Personal AI Assistant for GitHub, Google Drive, and Web",
    version="0.1.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


class Query(BaseModel):
    question: str
    sources: list[str] = ["github", "drive", "web"]


class Response(BaseModel):
    answer: str
    sources: list[dict]
    confidence: float


@app.get("/")
async def root():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "personAI",
        "version": "0.1.0"
    }


@app.get("/health")
async def health():
    """Detailed health check"""
    return {
        "status": "healthy",
        "read_only_mode": os.getenv("READ_ONLY_MODE", "true").lower() == "true",
        "available_sources": ["github", "drive", "web"]
    }


@app.post("/query", response_model=Response)
async def query(q: Query):
    """
    Query your personal data across GitHub, Google Drive, and the web
    
    Args:
        q: Query object containing question and desired sources
        
    Returns:
        Response with answer, sources, and confidence score
    """
    # TODO: Implement actual query logic
    return Response(
        answer="This is a placeholder response. Query logic to be implemented.",
        sources=[],
        confidence=0.0
    )


@app.post("/index/github")
async def index_github():
    """Index GitHub repositories for faster querying"""
    # TODO: Implement GitHub indexing
    return {"status": "not_implemented"}


@app.post("/index/drive")
async def index_drive():
    """Index Google Drive for faster querying"""
    # TODO: Implement Drive indexing
    return {"status": "not_implemented"}


if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", 8080))
    uvicorn.run(app, host="0.0.0.0", port=port)
