requirements.txt
# Core AI/LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.10.0

# Vector Store & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# GitHub Integration
PyGithub>=2.1.1
gitpython>=3.1.40

# Google APIs
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# Web Search & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.41.0

# FastAPI for API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
aiohttp>=3.9.0
tiktoken>=0.5.2

# Development
pytest>=7.4.0
black>=23.12.0
ruff>=0.1.11
